PEBA 2.0: remove Alert-storing, refactor API (#59)
We use a new tool written in Golang to store the honeypot alerts, and hence removed the parts of PEBA that were responsible for storage (sometimes called "PUT-Service"). PEBA only serves as egress API for data retrieval of already stored alerts. Also, there was some heavy refactoring done. This is already live in DTs backend.

Co-authored-by: vorband <vorband@users.noreply.github.com>